requirements.txt
absl-py==0.9.0
astor==0.8.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
iso-639==0.4.5
iso3166==1.0.1
isodate==0.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.2.0.34
pafy==0.5.5
Pillow==7.1.2
protobuf==3.12.2
pycryptodome==3.9.8
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
streamlink==1.4.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2020.6.16.1
zipp==3.1.0
